+++ lib/mad_flatter/options.rb
@@ -14,10 +14,16 @@ module MadFlatter
       def new(**options)
         immutable_struct_ex = ImmutableStructEx.new(**options) do
           def namespace?
+            return false unless respond_to? :namespace
+
             namespace || false
           end
 
           def metadata?
+            # This should never return false since
+            # the :metadata option is required.
+            return false unless respond_to? :metadata
+
             metadata || false
           end
         end
@@ -26,6 +32,8 @@ module MadFlatter
       end
 
       def with_defaults(options, defaults: DEFAULT_OPTIONS)
+        validate_options_type!(options: options) unless options.respond_to? :to_h
+
         new(**defaults.to_h.merge(options.to_h))
       end