+++ spec/mad_flatter_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe MadFlatter do
-  it "has a version number" do
-    expect(MadFlatter::VERSION).not_to be nil
-  end
-end